refactor(engine): extract crash recovery to src/daemon/startup-recovery.ts
readDaemonSessionState(), readAllDaemonSessions(), runStartupRecovery(),
countOrphanStepAdvances(), clearQueueIssueSidecars(), clearStrayTmpFiles(),
MAX_ORPHAN_AGE_MS, MAX_WORKTREE_ORPHAN_AGE_MS all extracted to a dedicated
crash-recovery module. These are a coherent concern (startup recovery from
crashes) that doesn't belong in the session execution file.

startup-recovery.ts imports runWorkflow from workflow-runner.ts as an injectable
default parameter -- this is safe because the call graph is acyclic at module
load time (no initialization-time circular dependency).

trigger-listener.ts updated to import runStartupRecovery directly from
startup-recovery.ts. workflow-runner.ts retains re-exports for backward compat.

workflow-runner.ts: 3,909 lines (original) -> 1,284 lines after this PR series.

378/378 test files, 5860 tests pass. Build clean.
