+++ src/trigger/trigger-listener.ts
@@ -29,7 +29,8 @@ import { TriggerRouter, type RunWorkflowFn } from './trigger-router.js';
 import { ActiveSessionSet } from '../daemon/active-sessions.js';
 import { loadWorkrailConfigFile, loadWorkspacesFromConfigFile } from '../config/config-file.js';
 import { NotificationService } from './notification-service.js';
-import { runWorkflow, runStartupRecovery } from '../daemon/workflow-runner.js';
+import { runWorkflow } from '../daemon/workflow-runner.js';
+import { runStartupRecovery } from '../daemon/startup-recovery.js';
 import type { WebhookEvent, WorkspaceConfig } from './types.js';
 import { asTriggerId } from './types.js';
 import type { DaemonEventEmitter } from '../daemon/daemon-events.js';
